Remove code to handle share intents from RemoteActivity to specific activities
Handling share intents in the remote activity was complicating the class's code, so moving it to separate specific activities is an improvement in its structure.
