--- app/src/main/java/org/xbmc/kore/ui/sections/remote/QueueActivity.java
@@ -1,12 +1,0 @@
-package org.xbmc.kore.ui.sections.remote;
-
-import android.content.Intent;
-
-public class QueueActivity extends RemoteActivity {
-
-    @Override
-    protected void handleStartIntent(Intent intent) {
-        handleStartIntent(intent, true);
-    }
-
-}
